refactor: move operation DTOs out of rest package
