+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/it/DeleteProcessInstanceOperationZeebeIT.java
@@ -10,8 +10,8 @@ package io.camunda.operate.it;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.operate.util.OperateZeebeAbstractIT;
+import io.camunda.operate.webapp.operation.dto.BatchOperationRequestDto;
 import io.camunda.operate.webapp.reader.BatchOperationReader;
-import io.camunda.operate.webapp.rest.dto.operation.BatchOperationRequestDto;
 import io.camunda.webapps.schema.entities.operation.BatchOperationEntity;
 import java.util.List;
 import org.junit.Before;
