+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/it/IncidentWithFailingOperationZeebeIT.java
@@ -20,8 +20,8 @@ import io.camunda.operate.util.OperateZeebeAbstractIT;
 import io.camunda.operate.util.OperationsManager;
 import io.camunda.operate.util.TestApplication;
 import io.camunda.operate.util.ZeebeTestUtil;
+import io.camunda.operate.webapp.operation.dto.CreateOperationRequestDto;
 import io.camunda.operate.webapp.rest.dto.incidents.IncidentDto;
-import io.camunda.operate.webapp.rest.dto.operation.CreateOperationRequestDto;
 import io.camunda.operate.webapp.zeebe.operation.ResolveIncidentHandler;
 import io.camunda.webapps.schema.entities.incident.ErrorType;
 import io.camunda.webapps.schema.entities.operation.OperationType;
