+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/webapp/zeebe/operation/process/modify/ModifyProcessInstanceOperationIT.java
@@ -23,10 +23,10 @@ import io.camunda.client.api.response.ModifyProcessInstanceResponse;
 import io.camunda.client.impl.CamundaClientFutureImpl;
 import io.camunda.operate.property.OperateProperties;
 import io.camunda.operate.util.j5templates.OperateZeebeSearchAbstractIT;
+import io.camunda.operate.webapp.operation.dto.ModifyProcessInstanceRequestDto;
+import io.camunda.operate.webapp.operation.dto.ModifyProcessInstanceRequestDto.Modification;
+import io.camunda.operate.webapp.operation.dto.ModifyProcessInstanceRequestDto.Modification.Type;
 import io.camunda.operate.webapp.reader.FlowNodeInstanceReader;
-import io.camunda.operate.webapp.rest.dto.operation.ModifyProcessInstanceRequestDto;
-import io.camunda.operate.webapp.rest.dto.operation.ModifyProcessInstanceRequestDto.Modification;
-import io.camunda.operate.webapp.rest.dto.operation.ModifyProcessInstanceRequestDto.Modification.Type;
 import io.camunda.webapps.schema.descriptors.template.BatchOperationTemplate;
 import io.camunda.webapps.schema.descriptors.template.OperationTemplate;
 import io.camunda.webapps.schema.entities.flownode.FlowNodeState;
