+++ operate/webapp/src/main/java/io/camunda/operate/webapp/opensearch/reader/OpensearchBatchOperationReader.java
@@ -19,8 +19,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.operate.conditions.OpensearchCondition;
 import io.camunda.operate.store.opensearch.client.sync.RichOpenSearchClient;
 import io.camunda.operate.util.CollectionUtil;
+import io.camunda.operate.webapp.operation.dto.BatchOperationRequestDto;
 import io.camunda.operate.webapp.reader.BatchOperationReader;
-import io.camunda.operate.webapp.rest.dto.operation.BatchOperationRequestDto;
 import io.camunda.operate.webapp.security.permission.PermissionsService;
 import io.camunda.webapps.schema.descriptors.template.BatchOperationTemplate;
 import io.camunda.webapps.schema.entities.operation.BatchOperationEntity;
