+++ operate/webapp/src/main/java/io/camunda/operate/webapp/operation/dto/BatchOperationDto.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.operate.webapp.rest.dto.operation;
+package io.camunda.operate.webapp.operation.dto;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.operate.webapp.rest.dto.listview.SortValuesWrapper;
