+++ operate/webapp/src/main/java/io/camunda/operate/webapp/transform/DataAggregator.java
@@ -8,7 +8,7 @@
 package io.camunda.operate.webapp.transform;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.camunda.operate.webapp.rest.dto.operation.BatchOperationDto;
+import io.camunda.operate.webapp.operation.dto.BatchOperationDto;
 import io.camunda.webapps.schema.entities.operation.BatchOperationEntity;
 import java.util.LinkedHashMap;
 import java.util.List;
