+++ operate/webapp/src/main/java/io/camunda/operate/webapp/zeebe/operation/MigrateProcessInstanceHandler.java
@@ -11,7 +11,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.client.api.command.MigrationPlan;
 import io.camunda.client.api.command.MigrationPlanBuilderImpl;
 import io.camunda.client.api.command.MigrationPlanImpl;
-import io.camunda.operate.webapp.rest.dto.operation.MigrationPlanDto;
+import io.camunda.operate.webapp.operation.dto.MigrationPlanDto;
 import io.camunda.spring.utils.ConditionalOnRdbmsDisabled;
 import io.camunda.webapps.schema.entities.operation.OperationEntity;
 import io.camunda.webapps.schema.entities.operation.OperationType;
